+++ example/lib/leyoujia/rent_page.dart
@@ -17,7 +17,7 @@ class RentPage extends StatefulWidget {
 }
 
 class _RentPageState extends State<RentPage> {
-  final _controller = DropdownSelectorController();
+  final DropdownSelectorController _controller = DropdownSelectorController();
   late final HouseRepository _repo;
   late final HouseFiltersRepository _filtersRepo;
   HouseFilter? _filter;
@@ -265,7 +265,7 @@ class _RentPageState extends State<RentPage> {
                 DropdownTab(
                   // tag: 'region',
                   label: l10n?.region ?? '',
-                  // labelGetter: (tabData, selected) {
+                  // labelLoader: (tabData, selected) {
                   //   // 可选：用户根据结果自定义标签
                   //   return '自定义标签';
                   // },
@@ -353,15 +353,15 @@ class _RentPageState extends State<RentPage> {
                 debugPrint('onHidden: ${tabData.label}');
               },
               onChanged: (tabData, selected) {
-
-                final result = DropdownSelectorResult(tabData: tabData, selected: selected);
+                final result = DropdownSelectorResult(
+                    tabData: tabData, selected: selected);
                 debugPrintLarge('onChanged: $result');
                 _handleSelectorChange(result);
                 _showSelectedResult(result);
               },
               onApplied: (tabData, selected) {
-
-                final result = DropdownSelectorResult(tabData: tabData, selected: selected);
+                final result = DropdownSelectorResult(
+                    tabData: tabData, selected: selected);
                 debugPrintLarge('onApplied: $result');
                 _handleSelectorApply(result);
                 if (result.tabIndex == 2) {